% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fsa_files.R
\name{process_fsa_files}
\alias{process_fsa_files}
\title{Process FSA Files}
\usage{
process_fsa_files(folder_path, ladder)
}
\arguments{
\item{folder_path}{A character string specifying the folder containing .fsa files.}

\item{ladder}{A vector or object representing the ladder information for processing.}
}
\value{
A Plotly object containing the interactive plot.
}
\description{
This function processes .fsa files from a specified folder, applies necessary transformations,
and generates an interactive plot for visualization.
}
